+++ views/admin/motion-type/_sections.php
@@ -42,7 +42,7 @@ $sName = 'sections[' . $sectionId . ']';
             ?>
             <label class="sectionTitle"><span class="sr-only"><?= Yii::t('admin', 'motion_section_name') ?></span>
                 <input type="text" name="<?= $sName ?>[title]" value="<?= Html::encode($section->title ?: '') ?>"
-                       required placeholder="Titel" class="form-control">
+                       required placeholder="<?= Yii::t('admin', 'motion_section_name_p') ?>" class="form-control">
             </label>
 
         </div>