+++ History.md
@@ -7,6 +7,7 @@
 - When editing a motion or amendment as admin, it is possible to set the status to "Obsoleted by another amendment" or "Obsoleted by another motion" and specify which one it is in a dropdown.
 - When creating a new consultation based on an existing one, it is now possible to choose if motion types, tags and/or user permissions are to be copied from the existing consultation or not.
 - When a list of organisations is specified in the consultation settings, this list is shown as a drop-down when creating / inviting new users.
+- The list of tags / topics can now be re-ordered and existing tags can be renamed.
 - It can now be set up so that the modified version proposed as part of a proposed procedure is shown inline as part of the motion.
 - Besides of exporting all motions as a ZIP-file containing single ODT files, a single ODT file containing all motion texts can now be exported.
 - The ODT export now also supports numbered lists.